Add Habitation. Change how we edit specific properties of single layers by using a new helper function.
