+++ helpers/cloneFromCatalogPath.js
@@ -1,5 +1,6 @@
 "use strict";
 const _ = require('lodash');
+const findInMembers = require('../helpers/findInMembers');
 
 /*
   json: Complete catalog file.
@@ -16,26 +17,6 @@ function cloneFromCatalogPath(json, path) {
     throw new Error("Could not find catalog item for path: " + path.join(","));
   }
   return _.cloneDeep(item);
-}
-
-function findInMembers(members, path) {
-  const first = path[0];
-  const rest = path.slice(1);
-
-  for (let i = 0; i < members.length; ++i) {
-    const item = members[i];
-    if (item.name === first) {
-      if (rest.length === 0) {
-        return item;
-      } else {
-        const match = findInMembers(item.members, rest);
-        if (match) {
-          return match;
-        }
-      }
-    }
-  }
-  return undefined;
 }
 
 module.exports = cloneFromCatalogPath;
